+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- fix and test regression forbiding to use Proxy models as GFK
+- Fix setting GFK parameter by a callable
+- Fix regression forbidding using Proxy models as GFK
 
 ## [1.20.1](https://pypi.org/project/model-bakery/1.20.1/)